created shared Problem class for HeatTransferProblem and HTransportProblem
